Abstracted CasterContext, CasterSettings, and changed GeometryRequester to GeometryStreamingService.
+ Created CasterContext from RemotePlayContext.
+ Created UnrealCasterContext, while I figure out how to abstract the video encoder.
+ Created CasterSettings to hold settings from RemotePlayMonitor.
+ Added ClientMessaging, but it doesn't do anything yet; intended for abstracting functionality of SessionComponent.
+ Created GeometryStreamingService, as a parent class of FGeometryStreamingService, to move more functionality to SimulCasterServer.
# Moved decomposition of streamable actors in level to RemotePlayMonitor from FGeometryStreamingService.
# GeometryStore stores hands as vector of pairs of <void*, avs::uid> from vector of IDs.

+++ SimulCasterServer/src/SimulCasterServer/NetworkPipeline.cpp
@@ -56,7 +56,7 @@ void NetworkPipeline::initialise(avs::Queue* ColorQueue, avs::Queue* DepthQueue,
 	}
 	geometryPipes[0]->sourceQueue = GeometryQueue;
 
-	int NumInputs = videoPipes.size() + geometryPipes.size();
+	size_t NumInputs = videoPipes.size() + geometryPipes.size();
 	if(!networkSink->configure(NumInputs, nullptr, localPort, remoteIP, remotePort))
 	{
 		std::cout << "Failed to configure network sink!\n";
